treewide: images: remove size specifications for root partitions
With current rootfs sizes of below 200 MiB there is no need to have a
512 MiB root partition, which will be a lot larger than the file system
it contains, and which will also all have to be written to the SD card.
Let genimage determine the partition size based on the file system image
it contains, thereby reducing the size of the generated disk images.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/20231103225253.1349209-4-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>