treewide: images: increase size of all qemu images to 1 GiB
qemu sets the size of the emulated disk as the size of the underlying
image file. We want to fit two rootfs partitions and one data partition
of 200 MiB each in there, which will need at least 600 MiB of disk
space. Round up to the next power of two, so qemu can work with it,
which results in 1 GiB, which the other qemu images already have too.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/20231103225253.1349209-5-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>