v7a: barebox: bootchooser: use GPT partition labels
Now that every image is in GPT format, we can use the devnodes that
include the partition label instead of the partition number to boot our
rootfs. These names are more readable and also more robust against
changing the order of the partitions. Make the change on all bareboxes
that didn't have that yet.

Signed-off-by: Roland Hieber <rhi@pengutronix.de>
Link: https://lore.pengutronix.de/20231103225253.1349209-11-rhi@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>