v7a: barebox-stm32mp1: pass root=PARTUUID=… to kernel
This partly reverts commit e0ca556db7983a99acc3330dc32cdae395e84003.

As Ahmad reports, the kernel device tree of STM32MP157C-EV1 has no
aliases at all, which is a precondition for enabling MMCBLKDEV_ROOTARG
in barebox. Restore the old state which detects the PARTUUID of the root
partition and passes it to the kernel.

Link: https://lore.distrokit.org/distrokit/9afb1d8a-f90d-7d74-781d-c33f6b7a2efb@pengutronix.de
Fixes: e0ca556db7983a99acc3 (2023-10-20, "v7a: barebox: pass root=mmcblkXpN instead of root=PARTUUID=… to kernel")
Reported-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Signed-off-by: Roland Hieber <rhi@pengutronix.de>