v7a: images: stm32mp: generate eMMC boot partition image as well
The DistroKit image is capable of booting both from SD-Card and eMMC
user area. For boot from eMMC boot partition, we can keep reusing the
same image for the user area, but need to place the boot firmware into
the boot partition as well. Add an extra image that does just that.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.pengutronix.de/20231213165014.1279779-2-a.fatoum@pengutronix.de
Signed-off-by: Robert Schwebel <r.schwebel@pengutronix.de>